pyproject.toml
[tool.poetry]
name = "ocrpack"
version = "0.1.0"
description = ""
authors = ["ArkarPhyo <arkarphyo.akp1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pipreqs = "^0.4.11"
albumentations = "^1.3.0"
fonttools = "^4.39.2"
lmdb = "^1.4.0"
numpy = "^1.24.2"
omegaconf = "^2.3.0"
pillow = "^9.4.0"
pyvips = "^2.2.1"
timm = "^0.6.13"
rich = "^13.3.3"
torch = "^2.0.0"
torchvision = "^0.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
